Move ShareFileView into its own module
